fix for cockroachdb

+++ src/routes.cpp
@@ -2,6 +2,7 @@
 #include "../include/token.h"
 #include "../include/db.h"
 #include "../include/regex.h"
+#include <cstdint>
 #include <string>
 
 //  Add User
@@ -39,7 +40,7 @@ std::shared_ptr<http_response> login_resource::render(const http_request& req) {
 
 std::shared_ptr<http_response> get_link_code_resource::render(const http_request& req) {
     std::string username = req.get_arg("token");
-    int id = get_user_id(username);
+    int64_t id = get_user_id(username);
     int code = generate_link_code(id);  
     
     return std::make_shared<string_response>(std::to_string(code), 200, "text/plain");
